rename example and make it better
